renamed zustand dir to store
